pascal triangle
The above code declares a function named printPascal which contain two nested loop. The outer loop starts from 1 and terminates at n and the inner loop starts from 1 to till the counter of outer loop .The variable D contains the series of numbers to be printed .The outer loop contains escape sequence “\n” which make it to the next line after each iteration.